refactor(test-cli): split cli contract support by responsibility
- separate assertions, process runner, and command result into dedicated files

- remove monolithic CliContractTestSupport container
