+++ tests/Ucli.Tests/Helpers/Cli/CommandExecutionResult.cs
@@ -1,0 +1,6 @@
+namespace MackySoft.Tests;
+
+internal readonly record struct CommandExecutionResult (
+    int ExitCode,
+    string StdOut,
+    string StdErr);
